added directory utils with two files hash_pass.py (hash password for save it in the file users.json) and user_file.py (this utils for open file with users and write file with users)
