+++ utils/hash_pass.py
@@ -1,0 +1,6 @@
+import hashlib
+
+
+# --- Утилиты ---
+def hash_password(password):
+    return hashlib.sha256(password.encode()).hexdigest()
