+++ utils/user_file.py
@@ -1,0 +1,16 @@
+import json
+
+
+# --- Загрузка пользователей ---
+def load_users():
+    try:
+        with open("users.json", "r") as f:
+            return json.load(f)
+    except FileNotFoundError:
+        return {}
+
+
+# --- Сохранение пользователей ---
+def save_users(users_db):
+    with open("users.json", "w") as f:
+        json.dump(users_db, f, indent=4)
